requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pyzbar
pillow
requests
python-telegram-bot
config
dataclasses_json
gspread